+++ src/_/collections/macros/hook.c
@@ -31,7 +31,7 @@ void add_hook_##name(return_type (*event) args_decl) {\
 }\
 \
 void run_hook_##name args_decl { \
-    for (int i = 0; i < functions_##name->size; i++) { \
+    for (size_t i = 0; i < functions_##name->size; i++) { \
         if (functions_##name->data[i].value != NULL) { \
             (*functions_##name->data[i].value) arg_names; \
         } \
