+++ src/flecs/util/post_update_loop.c
@@ -14,7 +14,7 @@ void add_to_post_update_loop(void (*event)(ecs_world_t *)) {
 }
 
 void run_post_update_loop(ecs_world_t *world) {
-    for (int i = 0; i < post_update_functions->size; i++) {
+    for (size_t i = 0; i < post_update_functions->size; i++) {
         if (post_update_functions->data[i].value != NULL) {
             (*post_update_functions->data[i].value)(world);
         }
