+++ src/flecs/util/update_loop.c
@@ -14,7 +14,7 @@ void add_to_update_loop(void (*event)(ecs_world_t *)) {
 }
 
 void run_update_loop(ecs_world_t *world) {
-    for (int i = 0; i < update_functions->size; i++) {
+    for (size_t i = 0; i < update_functions->size; i++) {
         if (update_functions->data[i].value != NULL) {
             (*update_functions->data[i].value)(world);
         }
