+++ src/rendering/basics3D/data/instance_render_command.c
@@ -9,7 +9,7 @@ byte has_mesh(
     InstanceRenderCommand_array_d* commands,
     ecs_entity_t mesh, int *index)
 {
-    for (int i = 0; i < commands->size; i++) {
+    for (size_t i = 0; i < commands->size; i++) {
         if (commands->data[i].mesh == mesh) {
             *index = i;
             return 1;
